requirements.txt
# Web Framework
streamlit==1.28.0

# Web Framework
streamlit==1.28.0

# Image Processing
Pillow==10.1.0
opencv-python==4.9.0.80

# AI/ML Models - Updated for Python 3.12
torch==2.2.0
torchvision==0.17.0
transformers==4.36.0
accelerate==0.25.0

# For Image Captioning/VQA (alternative to salesforce-lavis)
# Using BLIP through transformers instead
sentencepiece==0.1.99
protobuf==4.25.0

# Utilities
numpy==1.24.0
pandas==2.1.0
plotly==5.18.0
requests==2.31.0
tqdm==4.66.1